requirements.txt
aws_cdk.asset_node_proxy_agent_v5==2.0.67
